IMGui Docking & Static Libraries and ZERO Warnings
https://www.youtube.com/watch?v=lZuje-3iyVE
https://www.youtube.com/watch?v=TlvmnoDlrI0

+++ Hazel/src/Platform/Windows/WindowsWindow.cpp
@@ -135,7 +135,7 @@ namespace Hazel {
 		glfwSetScrollCallback(m_Window, [](GLFWwindow* window, double xOffset, double yOffset)
 		{
 			WindowData& data = *(WindowData*)glfwGetWindowUserPointer(window);
-			MouseScrolledEvent event(xOffset, yOffset);
+			MouseScrolledEvent event((float)xOffset, (float)yOffset);
 			data.EventCallback(event);
 		});
